Basic support for modern Puppet 3 type.
Include the self.instances and self.prefetch methods.
These are roughly the private entry type but intended
to populate the proptery_hash which also has to be
managed in create and destroy.

Due to the need to check the live system, the propery_hash
is avoided for actual system status checking.